Add DTO records and adjust numeric types
Add a set of request/response DTOs for aircraft, auth, crew, flight, delay and notification endpoints using Java records with Jakarta validation annotations (AircraftRequest/Response, LoginRequest, RegisterRequest, TokenResponse, CrewRequest/Response, CrewAssignmentRequest/Response, FlightRequest/Response, DelayRequest/Response, DelayResponse, NotificationResponse). Also change numeric fields to use Integer where appropriate: CrewMember.totalFlightHours (Long -> Integer) and FlightDelay.delayMinutes (Long -> Integer) to align types for counts/durations and simplify handling.

+++ src/main/java/az/azal/skyflow/crew/model/CrewMember.java
@@ -34,7 +34,7 @@ public class CrewMember extends BaseEntity {
 
     private LocalDateTime lastFlightEnd;
 
-    private Long totalFlightHours;
+    private Integer totalFlightHours;
 
     @Version
     private Long version;
